+++ rs/nns/cmc/CHANGELOG.md
@@ -10,6 +10,11 @@ here were moved from the adjacent `unreleased_changelog.md` file.
 
 INSERT NEW RELEASES HERE
 
+# Unreleased
+
+## Added
+
+- Added `environment_variables` to canister settings for `create_canister` and `notify_create_canister`
 
 # 2025-09-05: Proposal 138372